refactor: Fix a bunch of build errors and other issues
