+++ src/aux-server/aux-backend/server/.dev.env.json
@@ -1,0 +1,3 @@
+{
+    "DATABASE_URL": "postgresql://root@127.0.0.1:26257/defaultdb?sslmode=disable"
+}
