+++ src/aux-server/package.json
@@ -6,8 +6,8 @@
     "private": true,
     "scripts": {
         "start": "node aux-backend/server/dist/main.js",
-        "watch": "concurrently --names \"Server,Player\" \"npm:watch:server\" \"npm:watch:player\"",
-        "watch:secure": "concurrently --names \"Server,Player,Auth\" \"npm:watch:server\" \"npm:watch:player:secure\"",
+        "watch": "concurrently --names \"Server,Player,Auth\" \"npm:watch:server\" \"npm:watch:player\" \"npm:watch:auth\"",
+        "watch:secure": "concurrently --names \"Server,Player,Auth\" \"npm:watch:server\" \"npm:watch:player:secure\" \"npm:watch:auth:secure\"",
         "watch:server": "concurrently --names \"Build,Inst\" \"node ./script/dev-server.js\" \"npm:watch:inst\"",
         "watch:player": "vite serve ./aux-web/aux-player",
         "watch:auth": "vite serve ./aux-web/aux-auth",
@@ -136,10 +136,19 @@
         "vue-filepond": "^4.0.3",
         "vue-material": "1.0.0-beta-15",
         "vue-property-decorator": "^7.2.0",
-        "vue-router": "3.5.4"
+        "vue-router": "3.5.4",
+        "@prisma/client": "^4.14.1",
+        "aws-sdk": "^2.968.0",
+        "livekit-server-sdk": "1.0.2",
+        "luxon": "2.3.0",
+        "prisma": "^4.14.1",
+        "stripe": "11.13.0"
     },
     "devDependencies": {
         "@casual-simulation/aux-components": "^3.1.32",
         "schema-utils": "^3.0.0"
+    },
+    "prisma": {
+        "schema": "./aux-backend/schemas/auth.prisma"
     }
 }
