+++ src/aux-server/tsconfig.json
@@ -28,7 +28,9 @@
     "exclude": [
         "**/node_modules/**",
         "./aux-web/aux-player/sw.ts",
-        "./aux-web/aux-player/vite.config.ts"
+        "./aux-web/aux-player/vite.config.ts",
+        "./aux-web/aux-auth/sw.ts",
+        "./aux-web/aux-auth/vite.config.ts"
     ],
     "references": [
         { "path": "../aux-common" },
