+++ tsconfig.test.json
@@ -20,7 +20,6 @@
     "references": [
         { "path": "./src/crypto" },
         { "path": "./src/aux-server" },
-        { "path": "./src/aux-auth" },
         { "path": "./src/aux-common" },
         { "path": "./src/aux-components" },
         { "path": "./src/aux-vm" },
